% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nychex-package.R
\docType{package}
\name{nychex-package}
\alias{nychex}
\alias{nychex-package}
\title{nychex: Tessellated Hex Maps for New York City Geographies}
\description{
Provides sf objects containing hexagonal tile maps for New York City administrative geographies. Each polygon in the source geography is represented by a single hexagonal tile, preserving the overall spatial layout of the city.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kjhealy.github.io/nychex/}
  \item \url{https://github.com/kjhealy/nychex}
  \item Report bugs at \url{https://github.com/kjhealy/nychex/issues}
}

}
\author{
\strong{Maintainer}: Kieran Healy \email{kjhealy@gmail.com} (\href{https://orcid.org/0000-0001-9114-981X}{ORCID})

}
\keyword{internal}
